Add selected_node feature, and fix animation activation
Animation wasn't turning off when the tree reached its final state.
But also, it needed turned back on if a user drags an element.

Then I added a variable for selected_node, and added rendering for
it in the visualizers.  I also update the search blank with the
value of the selected node.