Replace broken murmurhash with actual NPM module, and add md5
My murmurhash approximation was showing really bad collisions
sometimes, and turned out it's because javascript doesn't perform
unsigned 32-bit math even when using bitwise operators.  Since the
actual correct version was so much more complicated, I just pulled
in the NPM module for it.  While I was at it, I pulled in a module
for md5 as well.