Overhaul API to primarily use path structs
I'm building the new type-agnostic API to use path structs to
reduce the number of parameters and overall complication.
Then there will be a type-specific API that can be declared
multiple times in multiple files, calling back to the base
API to prevent excessive code duplication.

I also allow the max_bits to be a macro, now, so that the full
library can be generated ahead of time and then the sizeof(size_t)
detected at compile time without exposing 64-bit types to a 32-bit
platform.  (or likewise for 16 bit)

Unfinished and untested, and relies on unpublished features of
CodeGen::Cpppp.