New API for generating routines with inlined compare functions
My new plan is to let the user declare their own function and then
set up the needed variables, then call the template to generate an
implementation using those variables.  This removes all the hassle
of declaring the parameter lists for the generated API and ending
up with redundant parameters.